requirements.txt
Flask==2.0.1
Flask-Dance==4.0.0
psycopg==3.1.8
python-dotenv==0.19.0
requests==2.26.0
bcrypt==3.2.0
gunicorn
